Move FastRecyclerViewActivity to separate module.
