--- app/src/main/java/com/example/kirchhoff/example/ui/recycler/fast/State.kt
@@ -1,3 +1,0 @@
-package com.example.kirchhoff.example.ui.recycler.fast
-
-class State(val name: String, val location: String)
