+++ fastrecyclerview/src/main/AndroidManifest.xml
@@ -1,0 +1,10 @@
+<manifest package="com.kirchhoff.fastrecyclerview"
+	xmlns:android="http://schemas.android.com/apk/res/android">
+
+	<application>
+		<activity
+			android:name=".FastRecyclerViewActivity"
+			android:label="@string/fast.recycler.view" />
+	</application>
+
+</manifest>
