+++ fastrecyclerview/src/main/java/com/kirchhoff/fastrecyclerview/State.kt
@@ -1,0 +1,3 @@
+package com.kirchhoff.fastrecyclerview
+
+class State(val name: String, val location: String)
