+++ fastrecyclerview/src/main/java/com/kirchhoff/fastrecyclerview/StateViewHolder.kt
@@ -1,10 +1,8 @@
-package com.example.kirchhoff.example.ui.recycler.fast
+package com.kirchhoff.fastrecyclerview
 
 import android.support.v7.widget.RecyclerView
 import android.view.View
 import android.widget.TextView
-
-import com.example.kirchhoff.example.R
 
 class StateViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
     var name: TextView = itemView.findViewById(R.id.name)
