+++ fastrecyclerview/src/main/res/layout/a_fast_recycler_view.xml
