+++ fastrecyclerview/src/main/res/layout/r_fast_recycler.xml
